requirements.txt
click==8.0.4
GitPython==3.1.27
PyYAML==6.0
tempfile==3.10.4